message relay file refactor
