+++ dev/Model/model_main.cpp
@@ -5,7 +5,7 @@
 #include <mutex>
 
 #include "Interfaces/types/state.h"
-#include "Messaging/system_messaging.h"
+#include "Messaging/message_relay.h"
 
 std::thread model_thread;
 std::mutex model_mutex;
