+++ dev/Model/model_main.h
@@ -1,7 +1,7 @@
 /*****************************************************************//**
  * \file   model_main.h
  * \brief  Main brains for the model. DO NOT INCLUDE. Include a model_interface instead
- * 
+ *
  * \author kason
  * \date   May 2021
  *********************************************************************/
@@ -14,13 +14,14 @@
 
 #include "Interfaces/types/model_command.h"
 #include "node/node.h"
-#include "Messaging/system_messaging.h"
+#include "Messaging/message_relay.h"
 
-/**
- * Main brain of the model. It keeps track of nodes and does actions such as looping, stepping through commands given from the Command_Interface, and sending nodes commands to run with devices.
- * \todo Include logging of model state.
- */
-namespace model {
+ /**
+  * Main brain of the model. It keeps track of nodes and does actions such as looping, stepping through commands given from the Command_Interface, and sending nodes commands to run with devices.
+  * \todo Include logging of model state.
+  */
+namespace model
+{
 	/**
 	 * .The string of the namespace name. This is for locating Messages.
 	 */
@@ -36,7 +37,7 @@ namespace model {
 	/**
 	 Actions to be completed on this step of the loop.
 	 */
-	extern Command_List step_actions; 
+	extern Command_List step_actions;
 
 	/**
 	 * \brief Get node from list of known nodes by ID
@@ -47,9 +48,9 @@ namespace model {
 
 	/**
 	 * \brief Create Node in models map.
-	 * 
+	 *
 	 * This alerts the model to another nodes existance. It can interact with that node as it sees fit.
-	 * 
+	 *
 	 * \param type Type of node based off of Enum
 	 * \param id Identifier of the node that model will refer to it as.
 	 */
@@ -57,7 +58,7 @@ namespace model {
 
 	/**
 	 * \brief A short circut for getting a device on a node using a node ID and device ID
-	 * 
+	 *
 	 * This is mainly used for testing devices, though should probably be removed.
 	 * \param label Node and device combination to access device.
 	 * \return Pointer to the called for device
@@ -66,16 +67,16 @@ namespace model {
 	Device* get_device(Device_Label);
 
 	/**
-	 * \brief Start up. 
-	 * 
-	 * Reset clock to 0 elapsed. Get node id from list of declared id's. Declare node operational to network and see who else is online. 
-	 * \todo network 
+	 * \brief Start up.
+	 *
+	 * Reset clock to 0 elapsed. Get node id from list of declared id's. Declare node operational to network and see who else is online.
+	 * \todo network
 	 */
 	void initalize();
 
 	/**
 	* \brief Run all commands lined up for this iteration of the loop.
-	* 
+	*
 	* This is the main function of a loop. If an exception is thrown, the command is thrown away and the exception gets rethrown.
 	* \todo Add details to the exception thrown
 	*/
@@ -83,14 +84,14 @@ namespace model {
 
 	/**
 	 * \brief Start shutting down the model.
-	 * 
+	 *
 	 * Shuts down the Model execution loop.
 	 */
 	void stop_loop();
 
 	/**
 	 * \brief Start up the Model loop.
-	 * 
+	 *
 	 * Spawns a thread for the Model to be run in.
 	 */
 	void start_loop();
@@ -112,6 +113,5 @@ namespace model {
 	 * \param command The command to run.
 	 */
 	void command_model(Model_Command command);
-
 }
 #endif
